requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-multipart==0.0.20
Pillow==11.1.0
pytesseract==0.3.13
sqlalchemy==2.0.36
aiosqlite==0.20.0
jinja2==3.1.5
python-dotenv==1.0.1
loguru==0.7.3
aiofiles==24.1.0
